SO MUCH
changed the  in every page back to normal.
updated title, h1 in header, and h2 in main.
updated the "read my introduction" text.
removed signature and privacy note from contract
added link in web115 subbullet to get back to the contract in the intro
darkened background to increase contrast (complies with WCAG now (i used the contrast checker))
i have also changed the white gradient in the header to be anchored to the top so that it does not go behind the white text.
changed fizzbuzzN links to fbN
changed "APs" to "AP classes" in the introduction
reduced visual spacing between header, main, and footer.
also reduced spacing above and below quote
intro -&gt; introduction